fixed some compiler warnings
--HG--
extra : convert_revision : svn%3Afdd8eb12-d10e-0410-9acb-85c331704f74/trunk%402351

+++ dlib/svm/kkmeans.h
@@ -271,7 +271,7 @@ namespace dlib
         // pick the first sample as one of the centers
         centers.push_back(samples[0]);
 
-        const long best_idx = samples.size() - samples.size()*percentile - 1;
+        const long best_idx = static_cast<long>(samples.size() - samples.size()*percentile - 1);
 
         // pick the next center
         for (long i = 0; i < num_centers-1; ++i)
